chore: thread and thread manager into separate file (#1336)
