+++ src/index.ts
@@ -4,6 +4,7 @@ export * from './client_state';
 export * from './channel';
 export * from './channel_state';
 export * from './thread';
+export * from './thread_manager';
 export * from './connection';
 export * from './events';
 export * from './moderation';
